+++ NEWS.md
@@ -1,5 +1,10 @@
 # ausvotesTR Development Version
 
+# 10 February 2023
+
+- Added `returns_url()` function, to add a URL to the specific return on the
+  AEC website.
+
 # 3 February 2023
 
 - Added a variable to `return_receipts_details` to indicate if an amount is